Created default layout for all sites
With the help of the front-end framework gethalfmoon.com I created
a html skeleton which should be used by all sites. The skeleton
uses the meta.html meta data and adds things like title, favicon
and a link to css.

Currently the css is loaded via a cdn and the favicon does not
exist.